requirements/requirements.txt
fastapi
pydantic
uvicorn
cassandra-driver